Move content in highlights.tsx to a new file key-features.tsx
